Hello, my name is Lucifer.  Welcome to GCC-Hell.  Let me give you a quick tour around the place.
On your left you see GCC2.  In GCC2, hash_map is included with
#include <hash_map> and lives in namespace std.

On your right you see GCC3.  In GCC3, hash_map is included with
#include <ext/hash_map> and lives in namespace __gnu_cxx.

Oh, and that little fellow in the middle?  Ignore him.  He's been running back
and forth for the past few hours mumbling shit about configure.in.

+++ p2psim/tapestry.C
@@ -1,4 +1,4 @@
-/* $Id: tapestry.C,v 1.2 2003/07/27 05:05:47 strib Exp $ */
+/* $Id: tapestry.C,v 1.3 2003/08/05 21:42:14 thomer Exp $ */
 
 #include "tapestry.h"
 #include "node.h"
@@ -7,15 +7,15 @@
 #include <iostream>
 #include <algorithm>
 #include <math.h>
-#include <vector.h>
+#include <vector>
 
 using namespace std;
 
 Tapestry::Tapestry(Node *n)
   : DHTProtocol(n),
     _base(16),
-    _bits_per_digit(log10(_base)/log10(2)),
-    _digits_per_id((uint)8*sizeof(GUID)/_bits_per_digit)
+    _bits_per_digit((uint) (log10(((double) _base))/log10((double) 2))),
+    _digits_per_id((uint) 8*sizeof(GUID)/_bits_per_digit)
 {
   joined = false;
   _my_id = get_id_from_ip(ip());
